+++ packages/terra-props-table/src/PropsTable.jsx
@@ -3,6 +3,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { parse } from 'react-docgen';
 import Markdown from 'terra-markdown';
+import classNames from 'classnames/bind';
+import styles from './PropsTable.scss';
+
+const cx = classNames.bind(styles);
 
 const propTypes = {
   /**
@@ -32,19 +36,19 @@ function determineType(type) {
     }
   } else if (typeName === 'arrayOf') {
     if (type.value.name === 'shape') {
-      typeName = <span> array of objects structured like: <pre> {formatShape(type.value.value)} </pre></span>;
+      typeName = <span> array of objects structured like: <pre className={cx('props-table-pre')}> {formatShape(type.value.value)} </pre></span>;
     } else {
       typeName = `array of ${type.value.name}s`;
     }
   } else if (typeName === 'union') {
     const options = type.value.map((option) => {
       const name = option.name === 'shape' ? (<span key={option.value}> an object structured like:
-        <pre> {formatShape(option.value)} </pre></span>) : (<span key={option.name}> {option.name}</span>);
+        <pre className={cx('props-table-pre')}> {formatShape(option.value)} </pre></span>) : (<span key={option.name}> {option.name}</span>);
       return name;
     });
     typeName = options.reduce((curr, next) => [curr, <span key={`${curr.value}-${next.value}`}> or </span>, next]);
   } else if (typeName === 'shape') {
-    typeName = <span> an object structured like: <pre> {formatShape(type.value)} </pre></span>;
+    typeName = <span> an object structured like: <pre className={cx('props-table-pre')}> {formatShape(type.value)} </pre></span>;
   }
 
   return typeName;
@@ -66,17 +70,19 @@ const PropsTable = ({ componentName, src, ...customProps }) => {
    */
   const componentProps = componentMetaData.props;
 
+  const tableRowClass = cx('prop-table-row');
+
   return (
     <div dir="ltr" className="markdown-body">
       <h2>{componentName} Props</h2>
       <table {...customProps}>
         <thead>
           <tr>
-            <th>Prop Name</th>
-            <th>Type</th>
-            <th>Is Required</th>
-            <th>Default Value</th>
-            <th>Description</th>
+            <th className={cx('prop-table-name')}>Prop Name</th>
+            <th className={cx('prop-table-type')}>Type</th>
+            <th className={cx('prop-table-required')}>Is Required</th>
+            <th className={cx('prop-table-default')}>Default Value</th>
+            <th className={cx('prop-table-description')}>Description</th>
           </tr>
         </thead>
         <tbody>
@@ -85,7 +91,7 @@ const PropsTable = ({ componentName, src, ...customProps }) => {
             const type = determineType(prop.type);
 
             return (
-              <tr key={key} style={{ fontSize: '90%' }}>
+              <tr className={tableRowClass} key={key} style={{ fontSize: '90%' }}>
                 <td style={{ fontWeight: 'bold' }}>{key}</td>
                 <td>{(prop.type ? type : '')}</td>
                 {(prop.required ?