+++ packages/terra-site/src/examples/form-input/examples/PlaceholderExample.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import Input from 'terra-form-input';
+
+const PlaceholderExample = () => (
+  <div>
+    <h4>Placeholder Text</h4>
+    <Input name="default placeholder input" placeholder="Placeholder Text" />
+  </div>
+);
+
+export default PlaceholderExample;