+++ packages/terra-props-table/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added styling to wrap long prop-type 'type' entries
+* Aligned Props Table entries to the top of their rows
 
 2.0.0 - (February 12, 2018)
 ------------------