+++ packages/terra-props-table/CHANGELOG.md
@@ -4,6 +4,12 @@ Changelog
 Unreleased
 ----------
 
+2.1.0 - (February 26, 2018)
+------------------
+### Added
+* Added styling to wrap long prop-type 'type' entries
+* Aligned Props Table entries to the top of their rows
+
 2.0.0 - (February 12, 2018)
 ------------------
 ### Changed